\name{test_function}
\alias{test_function}


\title{
This is my first function generated with documentation.
}
\description{

An X function

#' Well, this function is for writing stuff, I suppose that I need to write here the
#' pourpose of the function.
#' @param something Something to put
#' @keywords Tests, examples, many
#' @export
#' @examples
#' test_function()
}



\arguments{
  \item{x}{
Something
  }
  \item{\dots}{
Ignored.s
}
}
\author{
Juan Escamilla
}
