\name{presenceonly.BymCAR}
\alias{presenceonly.BymCAR}


\title{
Species Distribution Model for Presence-Only data
}
\description{
This function fits a Gaussian Markov Random Field for a given species and a given
sample. 
This is part of my thesis.
But in uses a Conditional Autoregressive Model jointly defined as i) The presence-only
data of a given species and ii) The Sample Effort defined by other type of data. 
Assumptions:
Both are Bernoulli observations.
based on (Besag, et. al 1993)
  
}

\usage{
\method{print}{CHANGE THIS LATER}(x, \dots)
}



\arguments{
  \item{x}{
    Developing...
}
  \item{\dots}{
Ignored.s
}
}
\author{
Juan Escamilla Mólgora
}
